부서, 직책, 직급 검색 및 수정 기능 구현

+++ src/main/java/site/currychoco/hrmanager/department/domain/DepartmentDto.java
@@ -13,7 +13,7 @@ public class DepartmentDto {
 
     private String deptCode;
     private String upperDeptCode;
-    private Integer headEmpNo;
+    private Long headEmpNo;
     private String deptName;
     private String deptNameEn;
     private int deptLevel;
